@@ -1,5 +1,5 @@
 import { Component, OnInit, ViewChild } from '@angular/core';
-import { FormControl, FormGroup, Validators } from '@angular/forms';
+import { UntypedFormControl, UntypedFormGroup, Validators } from '@angular/forms';
 import { MatTableDataSource } from '@angular/material/table';
 import { Router, ActivatedRoute } from '@angular/router';
 import { Pirate_Crew } from 'src/app/pirate-crew';
@@ -41,26 +41,26 @@ export class PiratesDetailComponent implements OnInit {
     'actions',
   ];
 
-  piratesDetailForm = new FormGroup({
-    crewId: new FormControl(null, [
+  piratesDetailForm = new UntypedFormGroup({
+    crewId: new UntypedFormControl(null, [
       Validators.required,
     ]),
   });
 
-  pirateCrewDetailForm = new FormGroup({
-    crewName: new FormControl(null, [
+  pirateCrewDetailForm = new UntypedFormGroup({
+    crewName: new UntypedFormControl(null, [
       Validators.required,
       Validators.minLength(5),
     ]),
-    captainName: new FormControl(null, [
+    captainName: new UntypedFormControl(null, [
       Validators.required,
       Validators.minLength(5),
     ]),
-    shipName: new FormControl(null, [
+    shipName: new UntypedFormControl(null, [
       Validators.required,
       Validators.minLength(5),
     ]),
-    totalMembers: new FormControl(null, [
+    totalMembers: new UntypedFormControl(null, [
       Validators.required
     ])
 
